made network more elaborate, fixed random num gen bug, creatures move up and down

+++ Assets/Scripts/ActionCreatorAbstract.cs
@@ -16,7 +16,7 @@ public abstract class ActionCreatorAbstract
         action.timeCost = cost;
     }
 
-    public void addResourceCost(string key, float value)
+    public void addResourceCost(string key, int value)
     {
         action.resourceCosts[key] = value;
     }
