+++ Assets/Scripts/Utility.cs
@@ -8,8 +8,6 @@ using Priority_Queue;
 public class Utility
 {
 
-
-
     public static System.Random rand = new System.Random();
     /*
     public static object getDeepCopy(object toCopy)
@@ -173,7 +171,7 @@ public class Utility
     public static Action getNewAction(Action oldAction)
     {
         Action newAction = oldAction.getShallowCopy();
-        Dictionary<string, float> dict = new Dictionary<string, float>();
+        Dictionary<string, int> dict = new Dictionary<string, int>();
         newAction.resourceCosts = dict;
         foreach (string key in oldAction.resourceCosts.Keys)
         {
